Move BeeInterface to game/characters/base/interfaces
