+++ frontend/models/game/characters/Healer.php
@@ -10,7 +10,7 @@ namespace frontend\models\game\characters;
 
 
 use frontend\models\game\characters\base\Bee;
-use frontend\models\game\base\BeeInterface;
+use frontend\models\game\characters\base\interfaces\BeeInterface;
 use frontend\models\game\pools\interfaces\HoneyPoolInterface;
 use frontend\models\game\characters\interfaces\HealerInterface;
