+++ frontend/models/game/characters/interfaces/HealerInterface.php
@@ -11,7 +11,7 @@ namespace frontend\models\game\characters\interfaces;
 
 use frontend\models\game\base\DoctorInterface;
 use frontend\models\game\base\SearchBeeInterface;
-use frontend\models\game\base\BeeInterface;
+use frontend\models\game\characters\base\interfaces\BeeInterface;
 
 interface HealerInterface extends BeeInterface, DoctorInterface, SearchBeeInterface
 {
