+++ frontend/models/game/characters/interfaces/WorkerInterface.php
@@ -10,7 +10,7 @@ namespace frontend\models\game\characters\interfaces;
 
 
 use frontend\models\game\base\HoneyMakerInterface;
-use frontend\models\game\base\BeeInterface;
+use frontend\models\game\characters\base\interfaces\BeeInterface;
 
 interface WorkerInterface extends BeeInterface, HoneyMakerInterface
 {
