+++ frontend/models/game/pools/interfaces/CharacterPoolInterface.php
@@ -9,7 +9,7 @@
 namespace frontend\models\game\pools\interfaces;
 
 
-use frontend\models\game\base\BeeInterface;
+use frontend\models\game\characters\base\interfaces\BeeInterface;
 use frontend\models\game\base\GetPlayerInterface;
 use frontend\models\game\base\KillAnyInterface;
 use frontend\models\game\base\SearchBeeInterface;
